Move configuration to a new file
